Quick fix: empty belongs_to ids crash.
Scenario: a User belongs to a Country and a Country has many Users.
In the User index we have table-plus-with-filters and you can filter by Country.
If you just click on Search, hobo-metasearch complains because there is no
such Country with id="nothing".

This is a quick fix, but I think we should enforce the input checking. What
about a regexp to ensure that this parameter is an integer for example?.